+++ app/controllers/api/v1/projects_controller.rb
@@ -29,4 +29,8 @@ class Api::V1::ProjectsController < Api::V1::BaseController
   # # GET /api/v1/projects/1-AUTHENTICATION_TOKEN/report-card.json
   # def report_card
   # end
+
+  # # GET /api/v1/projects/1-AUTHENTICATION_TOKEN/data-checks.json
+  # def data_checks
+  # end
 end