+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - Add support for reports that filter based on an expression, group by a
     specified date, and then list results by month and by site
   - Add support for "Report Card" report
+  - Add support for "Data Checks" report
 - **Adverse Event Changes**
   - Adverse event sheets can now be printed by reports, admins, and team members
 - **Export Changes**